+++ qgreenland/cli/__init__.py
@@ -3,6 +3,7 @@ import click
 from qgreenland.cli.cleanup import cleanup
 from qgreenland.cli.config_template import config_template
 from qgreenland.cli.fetch import fetch
+from qgreenland.cli.layers import layers
 from qgreenland.cli.provenance import provenance
 from qgreenland.cli.run import run
 
@@ -14,9 +15,10 @@ def cli():
 
 cli.add_command(cleanup)
 cli.add_command(config_template)
-cli.add_command(run)
 cli.add_command(fetch)
+cli.add_command(layers)
 cli.add_command(provenance)
+cli.add_command(run)
 
 
 if __name__ == '__main__':